+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Remove deprecated iDeal payment method
 * Apple Pay register_domains returns an ApplePayOptions object
 * Remove `unrecognized` status from Transaction, Subscription, and CreditCardVerification
+* Remove `GrantedPaymentInstrumentUpdate` kind from Webhook
 
 ## 3.59.0
 * Add `RefundAuthHardDeclined` and `RefundAuthSoftDeclined` to validation errors